+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Update default node-version for GitHub Action generator and fixes PG setup issues. ([@abhaynikam][])
 * Updates Bootstrap generator for supporting Bootstrap 5 with popper.js. ([@abhaynikam][])
 * Adds Overcommit generator for RuboCop. ([@coolprobn][])
+* Updates RuboCop generator to add rubocop-rake extension and support test frameworks ([@coolprobn][])
 * Adds Letter Opener generator. ([@coolprobn][])
 
 ## 0.11.0 (June 23rd, 2021)